proTES TLS ingress support
Add support for TLS termination for both the flower deployment
and proTES. Helm conditionals had to be used since the Ingress
objects definitions vary depending on the Ingress controller
which is on the target cluster. For OpenShift we use Route objects
and for Kubernetes we use Ingress objects.

The cluster type is defined under values.yaml, the possible values
are either 'openshift' or 'kubernetes'.

The documentation has been updated accordingly.