Travis CI config for proTES
Add support for continuous deployment of proTES in k8s clusters.
For this, one would need a service account in a namespace with
admin rights. For this repo we have created a dedicated namespace,
service account with correct access rights.

The authentication is handled via a token which is stored as a
secret in the service account's namespace.